+++ src/StdMeshers/StdMeshers_StartEndLength.cxx
@@ -153,11 +153,11 @@ istream & StdMeshers_StartEndLength::LoadFrom(istream & load)
 
   if (!isOK)
     load.clear(ios::badbit | load.rdstate());
-  
+
   isOK = (load >> intVal);
   if (isOK && intVal > 0) {
     _edgeIDs.reserve( intVal );
-    for (int i = 0; i < _edgeIDs.capacity() && isOK; i++) {
+    for ( size_t i = 0; i < _edgeIDs.capacity() && isOK; i++) {
       isOK = (load >> intVal);
       if ( isOK ) _edgeIDs.push_back( intVal );
     }
@@ -169,7 +169,7 @@ istream & StdMeshers_StartEndLength::LoadFrom(istream & load)
 
 //=============================================================================
 /*!
- *  
+ *
  */
 //=============================================================================
